requirements.txt
datasets
pandas
requests
python-dotenv
pystray
pillow
